requirements.txt
google-analytics-data==0.18.3
google-auth==2.33.0
google-auth-oauthlib==1.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
